+++ src/nmk_python/tests.py
@@ -5,6 +5,7 @@ Python tests builder
 import shutil
 import subprocess
 import sys
+from typing import cast
 
 from nmk.model.builder import NmkTaskBuilder
 from nmk.model.keys import NmkRootConfig
@@ -15,11 +16,12 @@ class PytestBuilder(NmkTaskBuilder):
     Python tests builder
     """
 
-    def build(self, pytest_args: dict[str, str]):
+    def build(self, pytest_args: dict[str, str | bool], ignore_failures: bool = False):  # type: ignore
         """
         Invoke pytest with specified options
 
         :param pytest_args: extra pytest command line args
+        :param ignore_failures: whether to ignore test failures
         """
 
         # Clean outputs
@@ -30,7 +32,7 @@ class PytestBuilder(NmkTaskBuilder):
                 p.unlink()
 
         # Compute extra args
-        args = []
+        args: list[str] = []
         for opt_k, opt_v in pytest_args.items():
             if isinstance(opt_v, bool):
                 if opt_v:
@@ -43,4 +45,4 @@ class PytestBuilder(NmkTaskBuilder):
         # Invoke pytest
         all_args = [sys.executable, "-m", "pytest"] + args
         self.logger.debug(f"Running subprocess: {' '.join(all_args)}")
-        subprocess.run(all_args, check=True, cwd=self.model.config[NmkRootConfig.PROJECT_DIR].value)
+        subprocess.run(all_args, check=not ignore_failures, cwd=cast(str, self.model.config[NmkRootConfig.PROJECT_DIR].value))